dept_add.jsp to create dept_add form at server side, policy_add.jsp to create policy_add form at server side, voucher_add.jsp to create voucher_add form at server side

+++ src/policy/Policy.java
@@ -148,7 +148,7 @@ public class Policy {
 	 * 
 	 * @param Integer
 	 */
-	public void setAmountPercent(int amountPercent) {
+	public void setAmountPercent(double amountPercent) {
 		this.amountPercent = amountPercent;
 	}
 	/**
